+++ CHANGELOG.md
@@ -3,5 +3,9 @@ All notable changes to the "jira-quick-info" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.2] 2024-07-21
+- Added total number of attachments and comments on tooltip
+- Renderer a mockup with issue ISSUE-1234
+
 ## [0.0.1] 2024-07-15
 - Initial release